refactor: Swap google-calendar-agent and inference-gateway service configurations in docker-compose.yaml
This repository is about the Google Calendar Agent, so I prioritize it at the top of the dependencies.

P.S the inference gateway is not a dependency, it's rather could be swapped with any inference provider directly that supports OpenAI - but for the sake of simplicity I will use it so I can easily swap and see the metrics.

Signed-off-by: Eden Reich <eden.reich@gmail.com>